refactoring signin
